services/ocr/requirements.txt
Flask==2.1.2
pytesseract==0.3.10
Pillow==9.4.0
psycopg2-binary==2.9.3
